Different approach to hostname management
Either update a hostname with all available instances or the last instance we heard launch.

Delete based on the route53 record will only happen when we are trying
to do "all available" and there are no instances left. (We want to
keep the "last instance to launch" in launch-before-terminate cases.)

Also AAAA records are set to track IPv6 addresses if we find them.

Also drop the default TTL to 60 to match that of the
[loadbalancer](https://docs.aws.amazon.com/elasticloadbalancing/latest/userguide/how-elastic-load-balancing-works.html#request-routing).